refactor: reorganize components and logic for badges
