--- src/pages/academy/grading/subcomponents/colors.ts
@@ -1,20 +1,0 @@
-const BADGE_COLORS = {
-  // assessment types
-  missions: 'indigo',
-  quests: 'emerald',
-  paths: 'sky',
-
-  // submission status
-  submitted: 'green',
-  attempting: 'yellow',
-  attempted: 'red',
-
-  // grading status
-  graded: 'green',
-  grading: 'yellow',
-  none: 'red'
-};
-
-export function badgeColor(label: string) {
-  return BADGE_COLORS[label.toLowerCase()] || 'gray';
-}
